+++ packages/@headlessui-vue/src/components/combobox/combobox.ts
@@ -742,6 +742,7 @@ export let Combobox = defineComponent({
                   hidden: true,
                   readOnly: true,
                   form,
+                  disabled,
                   name,
                   value,
                 })