+++ packages/@headlessui-vue/src/components/radio-group/radio-group.ts
@@ -264,6 +264,7 @@ export let RadioGroup = defineComponent({
                   hidden: true,
                   readOnly: true,
                   form,
+                  disabled,
                   name,
                   value,
                 })